Remove places debug panel && refactor css styles folder
