+++ src/components/landingComponents/Footer.css
@@ -1,4 +1,5 @@
 /* footer */
+
 .footer-landing {
   height: 150px;
   padding: 0 10rem;
@@ -7,12 +8,14 @@
   position: relative;
   color: #ededed;
 }
+
 .footer-landing div {
   flex: 1;
   display: flex;
   flex-direction: column;
   align-items: center;
 }
+
 #footer-background {
   height: 500px;
   width: 100%;
@@ -20,11 +23,12 @@
   bottom: 0;
   left: 0;
   right: 0;
-  background: url("../images/footer-dark-map.png");
+  background: url("../../images/footer-dark-map.png");
   background-size: cover;
   z-index: -1;
   box-shadow: 0 350px 500px 0px #0d0d0d inset;
 }
+
 .logo {
   height: 4rem;
 }
