+++ src/components/landingComponents/Footer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import logo from "../../images/white-logo.png";
-import "../../styles/Footer.css";
+import "./Footer.css";
 
 function Footer() {
   return (
