--- src/components/mapsComponents/Places.css
@@ -1,25 +1,0 @@
-.places-container {
-  position: absolute;
-  max-height: 20rem;
-  left: 0.5rem;
-  bottom: 0.5rem;
-  width: 24rem;
-  background-color: #fefefe;
-  border-radius: 4px;
-  box-shadow: 0 0 5px 0 #999999;
-  padding: 1rem;
-  overflow-y: auto;
-}
-.places-container p {
-  width: 100%;
-  overflow: hidden;
-  word-break: break-all;
-}
-.target-place {
-  margin-left: 16px;
-}
-@media (max-width: 500px) {
-  .places-container {
-    width: 16rem;
-  }
-}
